sys: kern: arch: rv32: move functions into respective module.
Some were missed with the initial move.
